Issue #13: Introduced 4 new Classes. Worked on MediaPlayer interaction.
Mostly, cries are now wrapped in the CentralMediaPlayer classes which for now fixes the amount of problems we might have had in terms of Android Media Player's state and adjusts that resourced to our needs.
Also, this commit introduces the PokedexManager for controlling context sensitive operations and interaction among all classes.