issue-#4-Scraping: Merged conflicts
Continued scrapper, descriptions now work. Working on hatchrate/catch/genderratios

Reorganized DB:

Created tables to represent hatchrate/catch/genderratios, removed the values from pokemontable and scraper file.

Also created the table @minicole indicated the caught table.

I ran the scrapper once, so the descriptions are filled up (But with new line chars, which is taken care off now, so it needs to be run again, but the code needs to be uncomentted as I was in the middle of testing for the others)