Issues #20 #12 #4: Stabilized application from constant crashing due to DB access
This was the result of merging this branch with master and branch issue-#12. There are several problems arising from database access for filling PokedexActivity including both structural DB changes and class implementation checks.
To stabilize the build, these were solved by commenting out several lines of code within PokedexFactory and commenting references within the UI.
However this does not fix the underlying problem which needs to be investigated and will require much refactoring in backend.