issue-#4-Scrapper:
Added a function to function_pokemonID to help find  a name from nationalID

Cleaned up some comments in pokemonTypes/Description

scraper_evolutions fills out the evolutions and their methods of pokemon. The specific order to run it in is mentioned in the header.
the databse was updated with this script, so just need to run move_database as usual.