Issue #12,#23: Massive additions to App Architecture
This commit includes substantial changes to the entire application, including the parallelization of database access and object creation as well as the instantiation of ProgressDialogs for UI feedback when building.

There are optimizations that can be done in terms of memory and object creation (forcing detailed object creation when asked might be an option). It is important to note the the Caught methods do not seem to be working correctly yet.